(std) refactor: break dependency on asserts testing module
