+++ std/encoding/toml.ts
@@ -1,6 +1,6 @@
 // Copyright 2018-2020 the Deno authors. All rights reserved. MIT license.
 import { deepAssign } from "../_util/deep_assign.ts";
-import { assert } from "../testing/asserts.ts";
+import { assert } from "../validation/assert.ts";
 
 class KeyValuePair {
   constructor(public key: string, public value: unknown) {}
