+++ std/http/cookie.ts
@@ -2,7 +2,7 @@
 // Structured similarly to Go's cookie.go
 // https://github.com/golang/go/blob/master/src/net/http/cookie.go
 import { ServerRequest, Response } from "./server.ts";
-import { assert } from "../testing/asserts.ts";
+import { assert } from "../validation/assert.ts";
 import { toIMF } from "../datetime/mod.ts";
 
 export interface Cookies {
