+++ std/http/file_server.ts
@@ -10,7 +10,7 @@ const { args, stat, readDir, open, exit } = Deno;
 import { posix, extname } from "../path/mod.ts";
 import { listenAndServe, ServerRequest, Response } from "./server.ts";
 import { parse } from "../flags/mod.ts";
-import { assert } from "../testing/asserts.ts";
+import { assert } from "../validation/assert.ts";
 
 interface EntryInfo {
   mode: string;
