+++ std/log/mod.ts
@@ -7,7 +7,7 @@ import {
   FileHandler,
   RotatingFileHandler,
 } from "./handlers.ts";
-import { assert } from "../testing/asserts.ts";
+import { assert } from "../validation/assert.ts";
 import { LevelName } from "./levels.ts";
 
 export { LogLevels } from "./levels.ts";
