+++ std/path/glob.ts
@@ -1,7 +1,7 @@
 import { SEP, SEP_PATTERN } from "./separator.ts";
 import { globrex } from "./_globrex.ts";
 import { join, normalize } from "./mod.ts";
-import { assert } from "../testing/asserts.ts";
+import { assert } from "../validation/assert.ts";
 
 export interface GlobOptions {
   extended?: boolean;
