+++ std/permissions/test.ts
@@ -1,7 +1,8 @@
 // Copyright 2018-2020 the Deno authors. All rights reserved. MIT license.
 
 import { grant, grantOrThrow } from "./mod.ts";
-import { assert, assertEquals } from "../testing/asserts.ts";
+import { assertEquals } from "../testing/asserts.ts";
+import { assert } from "../validation/assert.ts";
 
 const { test } = Deno;
