+++ std/validation/mod.ts
@@ -1,0 +1,5 @@
+// Copyright 2018-2020 the Deno authors. All rights reserved. MIT license.
+
+import * as assert from "./assert.ts";
+
+export { assert };
